Performance improvements
wordvectors::cosine:

This uses an intermediate vector (as before) to store the results and to
sort them. Instead of shrinking the output vector to the number of
wanted output words, it allocates a new vector and does the String
copying then. Since there are potentially multiple million words, the
number of String copies outweighs the cost of  an additional vector
allocation by far.

utils::norm:

Modify the vector in place, so that allocating a new vector is not
required anymore. This is possible without much changes, because the
vectors are actually always owned by the function which calls `norm()`.