WordVector::analogy: performance improvements
The previous code did copy all words (potentially multiple million
String copies). Afterwards, the results were sorted and the resulting
vector truncated to the requested length. Instead, all the Strings are
taken as a reference and after sorting, the number of actual hits are
collected into a vector with fresh string copies. As a nice side effect,
memory usage should be lower, too.